Se cambio el calculo para eficientizar la produccion y no cada maquina de automatizacion por separado. Se elimino codigo innecesario

+++ src/utils.h
@@ -11,7 +11,7 @@ using namespace std;
 void inicializarDB();
 bool validarExistenciaReceta(string nombreReceta);
 void crearReceta(string nombreReceta, int cantidadPorReceta);
-void calcularMaquinasNecesarias(string nombreReceta, int cantidad);
+void calcularMaquinasNecesarias(string nombreReceta, float cantidad);
 void imprimirMaquinasNecesarias();
 
 #endif
